Move ListemUser and UserDbContext to domain
