+++ Listem.API/Listem.API.csproj
@@ -27,6 +27,7 @@
     <ItemGroup>
       <Folder Include="Database\" />
       <Folder Include="Migrations\" />
+      <Folder Include="Repositories\" />
     </ItemGroup>
 
     <ItemGroup>
